Move `FilterHelper` and related classes to a shared location
These are currently being used by search, and will be used by the new list page, so move these to a better shared location
