+++ src/professions/interfaces/filter-input.interface.ts
@@ -1,0 +1,8 @@
+import { Industry } from '../../industries/industry.entity';
+import { Nation } from '../../nations/nation';
+
+export interface FilterInput {
+  nations: Nation[];
+  industries: Industry[];
+  keywords: string;
+}
