+++ src/professions/search/dto/filter.dto.ts
@@ -1,5 +1,5 @@
 export class FilterDto {
   keywords = '';
-  industries: string[] = [];
   nations: string[] = [];
+  industries: string[] = [];
 }
