+++ src/professions/search/search.controller.ts
@@ -6,8 +6,8 @@ import { IndustriesService } from '../../industries/industries.service';
 import { Nation } from '../../nations/nation';
 import { ProfessionsService } from '../professions.service';
 import { FilterDto } from './dto/filter.dto';
-import { FilterHelper } from './helpers/filter.helper';
-import { FilterInput } from './interfaces/filter-input.interface';
+import { FilterHelper } from '../helpers/filter.helper';
+import { FilterInput } from '../interfaces/filter-input.interface';
 import { IndexTemplate } from './interfaces/index-template.interface';
 import { SearchPresenter } from './search.presenter';
